rebase by hand 0305
