+++ pkg/mounter/proxy/server/alinas/driver.go
@@ -42,6 +42,10 @@ func (h *Driver) Mount(ctx context.Context, req *proxy.MountRequest) error {
 	return h.mounter.Mount(req.Source, req.Target, req.Fstype, options)
 }
 
+func (h *Driver) RotateToken(ctx context.Context, req *proxy.RotateTokenRequest) error {
+	return proxy.ErrNotImplemented(h.Name(), "rotateToken")
+}
+
 func (h *Driver) Init() {
 	go runCommandForever("aliyun-alinas-mount-watchdog")
 	go runCommandForever("aliyun-cpfs-mount-watchdog")
