+++ pkg/mounter/proxy/server/ossfs/utils.go
@@ -9,14 +9,6 @@ import (
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/mounter"
 )
 
-func writeFile(dir, fileName, contents string, perm os.FileMode) error {
-	file := filepath.Join(dir, fileName)
-	if err := os.Truncate(file, 0); err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	return os.WriteFile(file, []byte(contents), perm)
-}
-
 // rotateTokenFiles rotates (or initializes) token files
 func rotateTokenFiles(dir string, secrets map[string]string) (rotated bool, err error) {
 	if secrets == nil {
@@ -34,7 +26,7 @@ func rotateTokenFiles(dir string, secrets map[string]string) (rotated bool, err 
 			klog.Errorf("mkdirall tokendir failed %v", err)
 			return
 		}
-		err = writeFile(dir, key, val, 0o600)
+		err = os.WriteFile(filepath.Join(dir, key), []byte(val), 0o600)
 		if err != nil {
 			klog.Errorf("writeFile %s failed %v", key, err)
 			return
@@ -53,7 +45,7 @@ func prepareCredentialFiles(target string, secrets map[string]string) (file, dir
 	// fixed AKSK
 	hashDir := mounter.ComputeMountPathHash(target)
 	if passwd := secrets[OssfsPasswdFile]; passwd != "" {
-		err = writeFile(hashDir, OssfsPasswdFileName, passwd, 0o600)
+		err = os.WriteFile(filepath.Join(hashDir, OssfsPasswdFile), []byte(passwd), 0o600)
 		if err != nil {
 			return
 		}
